refactor vertex array construction
A new struct has been added: vertex_desc, holding generic data for a
vertex of a shape. An array of these define a shape entirely.

Specific transformations of the vertex_desc template such as
translating the positions to the correct cell and the second texture
coordinates according to the material, generate the vertices for a
vertex array