sz_Dict uses std::unordered_map, sz_Tag logs errors
The container associated with sz_Dict is an std::unordered_map
(probably a hash table; it depends on implementation). This simplifies
the loading code for all elements.

Added the sz_dict_lookup() function to operate on the new sz_Dict
implementation.

The get_* functions in sz_Tag write now a message to stderr explaining
what went wrong.

For this to work, the log_[info|warning|error] macros have been moved
to their own header file, so we don't have to include context.h just
for them.