box2, box3: add "cast" constructor and floor()
The cast constructor builds a box from a box of another type.

The floor function returns a box whose valiues are the floor() of the
original ones.

Combining these functions, the box for all cells in a given area can
be calculated in a short expression (as opposed to a variable
declaraton and six lines).