fix segflt when using "air" in a command
When the string "air" is passed as a block name, the shape is correctly
set to SHAPE_NONE, but the nmaterial is left undefined. Despite the
actual material value not mattering as fas as the game is concerned,
material values larger than 255 may be used as an index for the
block_traits table, producing a segmentation fault.