add List, Callback; register functions instead of entities
List is a doubly-linked list implementation where nodes must be
instances of a class that inherits from Node; this allows the Node to
automatically detach from the List on deletion.

Callback is a children of Node that has an std::function<void(void)> as
data; a callback list can be iterated an called, and callback objects
can be stored as std::unique_ptr to automatize the clean up process

During the update and render steps, instead of calling the update and
render functions for all entities, two callback lists are used;
entities may register methods for update and render by creating
callback objects and adding them to the lists on the Context and
Renderer classes. Callback methods can be member functions or simple
lambdas.