fix chunk update order when player is in intersection
Using std::map for a temporary chunk index by distance to the player is
a bad idea when the distances to several chunks can be the same; this
only happens when the player is in the intersection between two, so
I've needed a little while to catch up with it.

The solution is to use an std::multimap.