fix extra argument parsing, add option_parser namespace
The extra arguments with no trailing dash at the end of the list are
not interpreted as options but allowed to pass, and in the case of this
application used to determine the name of the world.

All option parsing classes and functions are now in the option_parser
namespace.