new crafting model
To check for recipes, first a CraftGrid must be built that maps
positions to actual inventory slots; currently only one inventory is
allowed per grid, but the framework for more is there.

More than one recipe may be matched to the same grid at once; in that
case, there must be a way for the player to chose the item to build.
Currently, the first possibility is automatically chosen.