CraftGrid: build from more than one inventory
A CraftGrid object can now be built form more than one inventory,
allowing for recipes larger than 3x3. None are yet implemented, and the
code is not yet in use by the PlayerEntity class.

For this to work, the box2 template class requires operator!=; it has
been implemented alongside operator==, both in box2 and box3.