Crafting in bench enabled by <shift>
instead of clicking at the side of the bench, crafting is benches
replaces the usual way to remove objects from the inventory, but
instead of taking the objects, the player may chose to take any of the
results of a recipe that involves the slot the cursor is pointing at.

Currently, this overrides the capability of taking back objects form
the bench inventory. A further update will add the pointed block to the
list of items that can be taken.