Fix block climbing
When the player moves against the side of a block, horizontal speed is
nullified in that direction, but vertical speed is reset so the player
will start climbing over the obstacle.

This fixes the player automagically going through roofs when moving
against walls in certain ways.

Also fixes the calculation if the top y coordinate during the collision
step, that didn't take into account the possibility of colliding with
several boxes with the same x/z coordinate but different tops.